+++ src/main/java/com/odeyalo/sonata/playlists/repository/r2dbc/R2dbcPlaylistOwnerRepository.java
@@ -1,4 +1,4 @@
-package com.odeyalo.sonata.playlists.repository;
+package com.odeyalo.sonata.playlists.repository.r2dbc;
 
 import com.odeyalo.sonata.playlists.entity.PlaylistOwnerEntity;
 import org.springframework.data.r2dbc.repository.R2dbcRepository;
