+++ src/main/java/com/odeyalo/sonata/playlists/repository/r2dbc/callback/write/SavePlaylistImageOnMissingAfterSaveCallback.java
@@ -4,7 +4,7 @@ import com.odeyalo.sonata.playlists.entity.PlaylistImage;
 import com.odeyalo.sonata.playlists.entity.ImageEntity;
 import com.odeyalo.sonata.playlists.entity.PlaylistEntity;
 import com.odeyalo.sonata.playlists.repository.PlaylistImagesRepository;
-import com.odeyalo.sonata.playlists.repository.R2dbcImageRepository;
+import com.odeyalo.sonata.playlists.repository.r2dbc.R2dbcImageRepository;
 import org.apache.commons.lang3.BooleanUtils;
 import org.jetbrains.annotations.NotNull;
 import org.reactivestreams.Publisher;
