+++ src/test/java/com/odeyalo/sonata/playlists/repository/r2dbc/R2dbcPlaylistRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.odeyalo.sonata.playlists.repository;
+package com.odeyalo.sonata.playlists.repository.r2dbc;
 
 import com.odeyalo.sonata.playlists.model.Images;
 import com.odeyalo.sonata.playlists.model.Playlist;
